moved wow_wrath behaviors
