--- behaviors/deathknight/frost.lua
@@ -1,1 +1,0 @@
-return {}
