--- behaviors/hunter/marksmanship.lua
@@ -1,1 +1,0 @@
-return {}
