--- behaviors/hunter/survival.lua
@@ -1,1 +1,0 @@
-return {}
