--- behaviors/mage/arcane.lua
@@ -1,1 +1,0 @@
-return {}
