--- behaviors/mage/fire.lua
@@ -1,1 +1,0 @@
-return {}
