--- behaviors/rogue/assassination.lua
@@ -1,1 +1,0 @@
-return {}
