--- behaviors/rogue/sublety.lua
@@ -1,1 +1,0 @@
-return {}
