--- behaviors/shaman/restoration.lua
@@ -1,1 +1,0 @@
-return {}
