--- behaviors/warlock/demonology.lua
@@ -1,1 +1,0 @@
-return {}
