--- behaviors/warlock/destruction.lua
@@ -1,1 +1,0 @@
-return {}
