+++ behaviors/wow_wrath/deathknight/frost.lua
@@ -1,0 +1,1 @@
+return {}
