+++ behaviors/wow_wrath/hunter/beastmastery.lua
