+++ behaviors/wow_wrath/hunter/marksmanship.lua
@@ -1,0 +1,1 @@
+return {}
