+++ behaviors/wow_wrath/hunter/survival.lua
@@ -1,0 +1,1 @@
+return {}
