+++ behaviors/wow_wrath/mage/arcane.lua
@@ -1,0 +1,1 @@
+return {}
