+++ behaviors/wow_wrath/mage/fire.lua
@@ -1,0 +1,1 @@
+return {}
