+++ behaviors/wow_wrath/paladin/common.lua
