+++ behaviors/wow_wrath/priest/common.lua
