+++ behaviors/wow_wrath/priest/shadow.lua
