+++ behaviors/wow_wrath/rogue/assassination.lua
@@ -1,0 +1,1 @@
+return {}
