+++ behaviors/wow_wrath/rogue/sublety.lua
@@ -1,0 +1,1 @@
+return {}
