+++ behaviors/wow_wrath/shaman/restoration.lua
@@ -1,0 +1,1 @@
+return {}
