+++ behaviors/wow_wrath/warlock/demonology.lua
@@ -1,0 +1,1 @@
+return {}
