+++ behaviors/wow_wrath/warlock/destruction.lua
@@ -1,0 +1,1 @@
+return {}
