+++ behaviors/wow_wrath/warrior/fury.lua
